#4: Добавлена загрузка таблицы ответов и разархивирование скачанных архивов.

+++ download_data.py
@@ -4,29 +4,56 @@ import requests
 from tqdm import tqdm
 import zipfile
 from zipfile import ZIP_DEFLATED, ZIP_LZMA
+import argparse
 
 root_url = "https://kpolyakov.spb.ru/download/"
-archive_names = ['24data.zip', '26data.zip', '27data.zip']
+archives_names = ['24data.zip', '26data.zip', '27data.zip']
+file_names = ['answers.csv'] + archives_names
 
-dirpath = Path('./data')
-if dirpath.exists() and dirpath.is_dir():
-    shutil.rmtree(dirpath)
-    print('Deleted old data files')
+if __name__ == '__main__':
+    print('\n Welcome to kpolyakov.spb.ru files downloader by Dmitry Kochetkov!\n')
 
-if not dirpath.exists():
-    dirpath.mkdir()
+    parser = argparse.ArgumentParser(description='Downloads problem files and answers from https://kpolyakov.spb.ru')
+    parser.add_argument('-f', metavar='force', action=argparse.BooleanOptionalAction, help='Force removal of data directory.')
+    args = parser.parse_args()
 
-for i, name in enumerate(archive_names):
-    print('[{} / {}] Downloading {}'.format(i+1, len(archive_names), name))
-    response = requests.get(root_url + name, stream=True)
+    dirpath = Path('./data')
+    if dirpath.exists() and dirpath.is_dir():
+        if args.f != True:
+            print('Directory ./data already exists. Use -f key to remove it.')
+            exit()
+        
+        shutil.rmtree(dirpath)
+        print('Erased ./data directory.')
 
-    with open("./data/" + name, "wb+") as handle:
-        for data in tqdm(response.iter_content()):
-            handle.write(data)
+    if not dirpath.exists():
+        dirpath.mkdir()
+        print('Created empty ./data directory')
+        print()
 
-    # with zipfile.ZipFile("./data/" + name, "r", ZIP_LZMA) as zip_ref:
-    #     unzip_directory_name = name.split('.')[0]
-    #     Path('./data/' + unzip_directory_name).mkdir()
-    #     zip_ref.extractall('./data/' + unzip_directory_name)
+    print('Downloading files...')
+    for i, name in enumerate(file_names):
+        print('[{} / {}] Downloading {}'.format(i+1, len(file_names), name))
+        response = requests.get(root_url + name, stream=True)
 
-print('Done')
+        with open("./data/" + name, "wb+") as handle:
+            for data in tqdm(response.iter_content()):
+                handle.write(data)
+
+    print('Download complete.')
+    print()
+    print('Extracting ZIP archives...')
+
+    for i, archive_name in enumerate(archives_names):
+        print('[{} / {}] Extracting {}'.format(i+1, len(file_names), archive_name))
+        
+        folder_name = archive_name.split('.')[0]
+        Path('./data/' + folder_name).mkdir()
+        with zipfile.ZipFile("./data/" + archive_name, "r", ZIP_DEFLATED) as f:
+            for name in tqdm(f.namelist()):
+                try:
+                    f.extract(name, path="./data/" + folder_name + "/")
+                except zipfile.BadZipfile:
+                    print('\nUnable to extract file', name)
+
+    print('Done')
